initial commit of circuitpy compatible code
main change needed to run on a circuitpy device are to get rid of uses of the warning module. Also changed time.monotonic to time.monotonic_ns to gain time resolution since sometimes microcontroller based controllers are sampling at high rates. I'll be considering this as I use the library more and may add a flag to the init to pick between the two